+++ config/urls.py
@@ -1,6 +1,9 @@
 from django.urls import include, path
 
-urlpatterns = [path("", include("drops.urls"))]
+urlpatterns = [
+    path("i18n/", include("django.conf.urls.i18n")),
+    path("", include("drops.urls")),
+]
 
 handler404 = "drops.views.not_found"
 handler500 = "drops.views.server_error"